+++ client/pages/post-gig/index.js
@@ -1,10 +1,18 @@
 import { useState } from "react";
-import { Box } from "../../styles/styles";
+import styled from 'styled-components'
+import Button from '../../styles/UI-Components/Button';
 
-//The number in the file names tell you which step in the form they are
-import GigBasics from "./1_GigBasics";
-import GigDetails from "./2_GigDetails";
-import GigSummary from "./3_GigSummary";
+import {
+  Box,
+  BoxTop,
+  GridList,
+  Title,
+  Title2,
+  ItemBox,
+  Input,
+  Selector,
+  Textarea
+} from '../../styles/styles';
 
 export default function CreateProject() {
   const [formData, setFormData] = useState({
@@ -16,6 +24,17 @@ export default function CreateProject() {
     gigamount: "",
     gigtimeframe: ""
   });
+
+  const {
+    gigname,
+    gigdescription,
+    gigwebsite,
+    gigrepo,
+    gigcategory,
+    gigreward,
+    gigamount,
+    gigtimeframe
+  } = formData;
 
   const [timeframeActive, setTimeframeActive] = useState(false);
 
@@ -61,7 +80,7 @@ export default function CreateProject() {
   const deleteCategory = (e) => {
     if (formData.gigcategory) {
       setFormData((prevCat) => {
-        const updatedSkills = prevCat.gigcategory.filter(
+        const updatedCategory = prevCat.gigcategory.filter(
           (category) => category !== e
         );
         return updatedCategory;
@@ -77,36 +96,205 @@ export default function CreateProject() {
   return (
       <Box>
           {basicsPage ? (
-            <GigBasics
-              goToDetails={goToDetails}
-              goBack={goBack}
-              formData={formData}
-              onChange={onChange}
-            />
+          <>
+            <BoxTop>
+              <Title>Gig Basics</Title>
+              <p><em>Let devs know what you are working on</em></p>
+            </BoxTop>
+
+              <InputSection>
+                <InputTitle>Write a name for your gig</InputTitle>
+                <Input
+                  placeholder='e.g., Smart Contract Developer for an NFT Game'
+                  name='gigname'
+                  value={gigname}
+                  onChange={e => onChange(e)}
+                />
+              <br/>
+                <InputTitle>Describe the gig clearly</InputTitle>
+                <Textarea
+                  minlength='30'
+                  placeholder='e.g., I am building an NFT game and need smart contract developers with experience in Solidity for a few weeks...'
+                  name='gigdescription'
+                  value={gigdescription}
+                  onChange={e => onChange(e)}
+                />
+              <br/>
+                <InputTitle>Does the project have a website? {'(optional)'}</InputTitle>
+                <Input
+                  placeholder='e.g., gigwebsite.com'
+                  name='gigwebsite'
+                  value={gigwebsite}
+                  onChange={e => onChange(e)}
+                />
+
+                <InputTitle>Does the project have a repo? {'(optional)'}</InputTitle>
+                <Input
+                  placeholder='e.g., github.com/gigname/gigrepo'
+                  name='gigrepo'
+                  value={gigrepo}
+                  onChange={e => onChange(e)}
+                />
+              </InputSection>
+
+            <ButtonSection>
+              <Button
+                onClick={goToDetails}
+                styling='positive'
+              >Continue</Button>
+              <Button
+                onClick={goBack}
+                styling='negative'
+              >Cancel</Button>
+          </ButtonSection>
+        </>
           ) : null}
 
           {detailsPage ? (
-            <GigDetails
-              goToBasics={goToBasics}
-              goToSummary={goToSummary}
-              formData={formData}
-              setFormData={setFormData}
-              updateGigCategory={updateGigCategory}
-              deleteCategory={deleteCategory}
-              timeframeActive={timeframeActive}
-              addTimeframe={addTimeframe}
-              onChange={onChange}
-            />
+            <>
+              <BoxTop>
+                <h1>Gig Details</h1>
+                <em>Define your project timeframe, category, and reward</em>
+              </BoxTop>
+
+                <InputSection>
+                  <InputTitle>Pick category</InputTitle>
+                  <CategorySection>
+
+                  </CategorySection>
+
+                  <InputTitle>Explain reward method {'(optional)'}</InputTitle>
+                  <select
+                    name='gigreward'
+                    value={gigreward}
+                    onChange={e => onChange(e)}
+                    >
+                    <option value="" disabled selected hidden>Token</option>
+                    <option value="ETH">ETH</option>
+                    <option value="SOL">SOL</option>
+                    <option value="BTC">BTC</option>
+                    <option value="USDT">USDT</option>
+                  </select>
+
+                  <Input
+                    placeholder='Amount'
+                    name='gigamount'
+                    value={gigamount}
+                    onChange={e => onChange(e)}
+                    type='number'
+                  />
+
+                  <div className='cp-input-checkbox'>
+                    <Input
+                    type='checkbox'
+                    onClick={addTimeframe}
+                    /> <label>Specific timeframe?</label>
+                  </div>
+                  {timeframeActive &&
+                    <Selector>
+                    <option value="" disabled selected hidden>Duration</option>
+                    <option value="0-30 Days">0-30 Days</option>
+                    <option value="1-3 Months">1-3 Months</option>
+                    <option value="3-6 Months">3-6 Months</option>
+                    <option value="6-12 Months">6-12 Months</option>
+                    <option value="+12 Months">+12 Months</option>
+                  </Selector>
+                  }
+
+                </InputSection>
+              <ButtonSection>
+                <Button styling='positive'
+                  onClick={goToSummary}
+                >Continue</Button>
+                <Button styling='negative'
+                  onClick={goToBasics}
+                >Back</Button>
+            </ButtonSection>
+          </>
           ) : null}
 
           {summaryPage ? (
-            <GigSummary
-              formData={formData}
-              goToDetails={goToDetails}
-              goToBasics={goToBasics}
-              createGig={createGig}
-            />
+            <>
+            <BoxTop>
+              <h1>Job Summary</h1>
+              <h2><em>Check that everything is correct {'(click edit if not)'}</em></h2>
+            </BoxTop>
+
+                <InputSection>
+                  <h2>Gig Basics</h2>
+                  <Button
+                    onClick={goToBasics}
+                    styling='normal'
+                  >Edit</Button>
+                </InputSection>
+                <br/>
+                <InputSection>
+                  <InputTitle>Gig Name:</InputTitle>{' '}<p>{gigname}</p>
+                </InputSection>
+
+                <br/>
+                <InputSection>
+                  <InputTitle>Gig Description:</InputTitle>{' '}{gigdescription}
+                </InputSection>
+                <br/>
+                  <InputSection>
+                    <InputTitle>Gig Website:</InputTitle>{' '}{gigwebsite}
+                  </InputSection>
+                <br/>
+
+              <InputSection>
+                <h2>Gig Details</h2>
+                <Button
+                  onClick={goToDetails}
+                  styling='normal'
+                >Edit</Button>
+              </InputSection>
+
+                <InputTitle>Category:</InputTitle>
+                <CategorySection>
+
+                </CategorySection>
+              <br/>
+              {formData && gigreward && gigamount ? (
+              <div>
+                <InputTitle>Gig Reward:</InputTitle>{' '} <p>{gigamount}{' '}{gigreward}</p>
+              </div>) : null}
+              <br/>
+              {formData && gigtimeframe ? (
+                <div>
+              <InputTitle>Gig Timeframe:</InputTitle>{' '}<p>{gigtimeframe}</p>
+              </div>) : null}
+
+
+              <ButtonSection>
+                <Button
+                  styling='positive'
+                  onClick={createGig}
+                >Create Gig</Button>
+
+                <Button
+                  onClick={goToDetails}
+                  styling='negative'
+                >Back</Button>
+              </ButtonSection>
+            </>
           ) : null}
       </Box>
   );
 }
+
+const InputSection = styled(ItemBox)`
+  box-shadow: 0 0 0 0;
+`;
+
+const InputTitle = styled(Title2)`
+  font-size: 1rem;
+  text-align: left;
+`;
+
+const ButtonSection = styled(ItemBox)`
+  box-shadow: 0 0 0 0;
+`;
+
+const CategorySection = styled(GridList)`
+`;
